airflow_pipeline/requirements.txt
# NBA Sportsbook Pipeline Dependencies
requests==2.31.0
pandas==2.1.4

# Airflow
apache-airflow==2.8.0

# Google Drive API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

python-dateutil==2.8.2
